+++ libs/__init__.py
@@ -1,2 +1,2 @@
-__version_info__ = ('1', '8', '14')
+__version_info__ = ('1', '8', '15')
 __version__ = '.'.join(__version_info__)